Replaced custom popper with popper
